requirements.txt
click>=8.1.7
rich>=13.7.0
pyyaml>=6.0.1
psutil>=5.9.8
python-dateutil>=2.8.2
watchdog>=4.0.0
